Created all the different acts, changed the start script
We now have 3 different acts! All possible to go to once you start up a new game.

Renamed the "script" file to "start-script" to make it more clear what's it used for.

I also added a "remove" character that needs to be removed before uploading. It's only there to make it clear you are in a different act.