Implemented blurry and shaking, can be seen through the menu from start-script
I edited the bg forest night blurry picture to make it look blurry, can be done for any picture of course but this is just an example.

In the script "blurryShake" file there's all the code with explanations on how it works.

Also added a menu option to test this out