PNG crushed all PNGs with OptiPNG and PNG crushed all backgrounds and CGs with PNGout. Changed script slightly.
PNG crushed all PNGs with OptiPNG and PNG crushed all backgrounds and CGs with PNGout.

PNGout is better at lossless PNG compression but has to be done per file, so I just did it on the BGs and CGs and base sprites. I also changed some wording to be more British/Aussie.