WebP update for Android build. Also added Android downloading screen.
Background and CGs, save for the scrolling CG in act 3, don't actually refer to any file extension, so I replaced them to save 80 MB or so on the Android build.
I also added a custom downloading screen with pre-existing art to replace the default Ren'py one with the girl on the Android rocket.

The new build is already in review on the Play Store.